split tests between two classes
