+++ .circleci/config.yml
@@ -211,7 +211,7 @@ jobs:
             -   run: # collect results
                     name: Save test results
                     command: |
-                        mkdir -p ~/rm test-results/junit/
+                        mkdir -p ~/test-results/junit/
                         find . -type f -regex ".*/build/test-results/.*xml" -exec cp {} ~/test-results/junit/ \;
                     when: always
             -   run: # upload coverage
