+++ .idea/kotlinScripting.xml
@@ -1,6 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
   <component name="KotlinScriptingSettings">
+    <option name="isAutoReloadEnabled" value="true" />
     <option name="suppressDefinitionsCheck" value="true" />
   </component>
 </project>
