+++ plugin/src/functionalTest/kotlin/com/klarna/gradle/reactnative/GradleReactNativePluginFunctionalTest.kt
@@ -1,19 +1,10 @@
-/*
- * This Kotlin source file was generated by the Gradle 'init' task.
- */
+/* */
 package com.klarna.gradle.reactnative
 
-import org.gradle.testkit.runner.GradleRunner
-import org.gradle.testkit.runner.InvalidPluginMetadataException
-import org.gradle.testkit.runner.internal.PluginUnderTestMetadataReading
-import org.gradle.util.GUtil
-import org.junit.Rule
-import org.junit.rules.TestName
-import java.io.File
 import kotlin.test.BeforeTest
+import org.gradle.testkit.runner.GradleRunner
 import kotlin.test.Test
 import kotlin.test.assertTrue
-import kotlin.text.RegexOption.IGNORE_CASE
 
 //region Constants
 const val DOLLAR = "\$"
@@ -86,78 +77,10 @@ plugins {
 //endregion
 
 /** Functional tests. Try to run plugin in different modes. */
-class GradleReactNativePluginFunctionalTest {
-    private var projectDir: File
-    private val jacocoDir: File
-    private val metadataDir: File
-
-    @Rule
-    @JvmField
-    public val testName: TestName = TestName()
-
-    init {
-        val pwd = File("").absolutePath
-        val projectRootDirName = "/gradle-react-native/plugin"
-        if (!pwd.contains(projectRootDirName)) {
-            throw AssertionError("Wrong working directory set for functional tests.")
-        }
-        projectDir = File("build/functionalTest")
-        jacocoDir = File("build/jacoco/functionalTest")
-        metadataDir = File("build/pluginUnderTestMetadata")
-    }
-
+class GradleReactNativePluginFunctionalTest() : CommonFunctionalTest() {
     @BeforeTest
-    fun initializeDirectory() {
-        // make test name safe for using as a file name
-        val testName = testName.methodName.replace(
-            "[^a-z0-9.-]".toRegex(IGNORE_CASE), "_"
-        )
-
-        projectDir = File("build/functionalTest/$testName")
-        projectDir.mkdirs()
-
-        // make empty files
-        projectDir.resolve("settings.gradle").writeText("$ANDROID_PLUGIN")
-        projectDir.resolve("build.gradle").writeText("")
-
-        // runtime jacoco attaching
-        val expandedDir = "build/tmp/expandedArchives"
-        val jacocoVer = "org.jacoco.agent-0.8.4.jar"
-        val jacocoRuntime: File = File("$expandedDir/$jacocoVer/jacocoagent.jar")
-        val javaAgent = "-javaagent:${jacocoRuntime.absolutePath}" +
-            "=destfile=${jacocoDir.absolutePath}/$testName.exec"
-        val memory = "-Xmx64m -Xms64m" +
-            " -Dkotlin.daemon.jvm.options=\"-Xmx64m\"" +
-            " -Dkotlin.compiler.execution.strategy=\"in-process\""
-        projectDir.resolve("gradle.properties").writeText(
-            """
-            # method=${this.testName.methodName}
-            org.gradle.caching=false
-            org.gradle.daemon=false
-            org.gradle.jvmargs=$javaAgent $memory -Dfile.encoding=UTF-8
-            """.trimIndent()
-        )
-    }
-
-    /** Helper that should solve classpath loading for Unit tests running in IDE. */
-    private fun loadClasspath(): List<File> {
-        try {
-            return PluginUnderTestMetadataReading.readImplementationClasspath()
-        } catch (ignored: InvalidPluginMetadataException) {
-            /* we cannot extract `plugin-under-test-metadata.properties` file from resources. */
-        }
-
-        with(File(metadataDir, "plugin-under-test-metadata.properties")) {
-            if (!exists()) {
-                throw AssertionError("./gradlew :plugin:pluginUnderTestMetadata not called.")
-            }
-            val properties = GUtil.loadProperties(this)
-
-            return PluginUnderTestMetadataReading.readImplementationClasspath(
-                absolutePath,
-                properties
-            )
-        }
+    override fun initializeDirectory() {
+        super.initializeDirectory()
     }
 
     @Test
@@ -180,56 +103,6 @@ class GradleReactNativePluginFunctionalTest {
             .buildAndFail()
 
         assertTrue(result.output.contains(GradleReactNativePlugin.EXCEPTION_NO_ANDROID_PLUGIN))
-    }
-
-    @Test
-    fun `can run compile task`() {
-        // Setup the test build
-        projectDir.resolve("build.gradle").writeText(
-            """
-            $GRADLE_DEPENDENCIES
-            $GRADLE_PLUGINS
-            android {
-                $ANDROID_SECTION
-            }
-            """.trimIndent()
-        )
-
-        // Run the build
-        val result = GradleRunner.create()
-            .forwardOutput()
-            .withPluginClasspath(loadClasspath())
-            .withProjectDir(projectDir)
-            .withArguments(CompileRnBundleTask.NAME)
-            .build()
-
-        // Verify the result
-        assertTrue(result.output.contains(CompileRnBundleTask.DUMMY))
-    }
-
-    @Test
-    fun `can run copy task`() {
-        // Setup the test build
-        projectDir.resolve("build.gradle").writeText(
-            """
-            $GRADLE_DEPENDENCIES
-            $GRADLE_PLUGINS
-            android {
-                $ANDROID_SECTION
-            }
-            """.trimIndent()
-        )
-
-        // Run the build
-        val result = GradleRunner.create()
-            .forwardOutput()
-            .withPluginClasspath(loadClasspath())
-            .withProjectDir(projectDir)
-            .withArguments(CopyRnBundleTask.NAME)
-            .build()
-
-        // Verify the result
-        assertTrue(result.output.contains(CopyRnBundleTask.DUMMY))
     }
 
     @Test
